fix crash of order list

+++ modulePublic/moduleCommon/src/main/java/com/einyun/app/common/repository/DatabaseRepo.java
@@ -59,13 +59,13 @@ public abstract class DatabaseRepo<T> implements IDatabaseRepo<T>{
 
     @Override
     public void persistence(List<T> rows,String userId,int listType, int dataType){
-//        if (dataType == DATA_TYPE_INIT) {//初始化清空数据
+        if (dataType == DATA_TYPE_INIT) {//初始化清空数据
             if(listType<0){
                 deleteAll(userId);
             }else{
                 deleteAll(userId,listType);
             }
-//        }
+        }
         insert(rows);//追加数据
     }
 }
